added rub in adapter

+++ app/src/main/java/com/ozontech/homework2/domain/mappers/ProductInListMapper.kt
@@ -8,6 +8,6 @@ fun ProductInListDO.toVO() = ProductInListVO(
     guid,
     image,
     name,
-    price,
+    price.toInt(),
     rating.toFloat()
 )
